# app/server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  pkg.ErrCode:
    enum:
    - 0
    - 1001
    - 1002
    - 1003
    - 1010
    - 1011
    - 1012
    - 1013
    - 1014
    - 1015
    - 1016
    - 2001
    - 3001
    - 3002
    - 5001
    - 6001
    - 6002
    - 7001
    - 7002
    - 7003
    - 7004
    - 8001
    - 8002
    - 8003
    - 8004
    - 8005
    - 8006
    - 8007
    - 8008
    - 8009
    - 8010
    - 8101
    - 8102
    - 8103
    - 8104
    type: integer
    x-enum-varnames:
    - SUCCESS
    - TOKEN_NOT_EXIST
    - TOKEN_WRONG
    - TOKEN_EXPIRED
    - INVALID_PASSWORD
    - USER_NOT_FOUND
    - USERNAME_EXISTS
    - FORBIDDEN
    - TOKEN_REVOKED
    - USER_BANNED
    - USER_MUTED
    - INVALID_PARAMS
    - NOT_FOUND
    - ALREADY_EXISTS
    - INTERNAL_ERROR
    - SFU_NOT_CONFIGURED
    - SFU_ERROR
    - OAUTH_PROVIDER_NOT_FOUND
    - OAUTH_PROVIDER_DISABLED
    - OAUTH_TOKEN_EXCHANGE_FAILED
    - OAUTH_GET_USER_FAILED
    - EMAIL_ALREADY_EXISTS
    - EMAIL_CODE_INVALID
    - EMAIL_CODE_EXPIRED
    - EMAIL_CODE_ALREADY_USED
    - EMAIL_SEND_TOO_FREQUENT
    - EMAIL_SEND_FAILED
    - EMAIL_NOT_VERIFIED
    - EMAIL_CODE_MAX_ATTEMPTS
    - EMAIL_NOT_CONFIGURED
    - PASSWORD_RESET_DISABLED
    - STORAGE_NOT_CONFIGURED
    - STORAGE_ERROR
    - STORAGE_FILE_TOO_LARGE
    - STORAGE_FILE_TYPE_NOT_ALLOWED
  pkg.Response:
    properties:
      code:
        $ref: '#/definitions/pkg.ErrCode'
      data: {}
      msg:
        type: string
    type: object
...
host: localhost:8998
info:
  contact:
    name: NoelOrin
    url: https://github.com/NoelOrin
  description: GoRTC - WebRTC 服务器端 API 接口文档
  license:
    name: MIT 许可证
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/NoelOrin/GoRTC
  title: GoRTC API
  version: "1.0"
paths:
...
  /user/delete:
    post:
      consumes:
      - application/json
      description: 根据用户 ID 删除用户
      parameters:
      - description: 用户 ID
        in: body
        name: request
        required: true
        schema:
          properties:
            id:
              type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg.Response'
      security:
      - BearerAuth: []
      summary: 删除用户
      tags:
      - 用户
  /user/get:
    post:
      consumes:
      - application/json
      description: 通过用户 ID 获取用户信息
      parameters:
      - description: 用户 ID
        in: body
        name: request
        required: true
        schema:
          properties:
            id:
              type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg.Response'
      security:
      - BearerAuth: []
      summary: 根据 ID 获取用户
      tags:
      - 用户
  /user/list:
    post:
      consumes:
      - application/json
      description: 分页获取用户列表
      parameters:
      - description: 分页参数
        in: body
        name: request
        required: true
        schema:
          properties:
            page:
              type: integer
            page_size:
              type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg.Response'
      security:
      - BearerAuth: []
      summary: 获取用户列表
      tags:
      - 用户
  /user/profile:
    post:
      description: 获取当前已认证用户的资料信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg.Response'
      security:
      - BearerAuth: []
      summary: 获取用户资料
      tags:
      - 用户
  /user/update-profile:
    post:
      consumes:
      - application/json
      description: 更新当前已认证用户的显示名称和头像
      parameters:
      - description: 用户资料
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg.Response'
      security:
      - BearerAuth: []
      summary: 更新当前用户资料
      tags:
      - 用户
  /user/update-role:
    post:
      consumes:
      - application/json
      description: 管理员更新用户角色（admin/user）
      parameters:
      - description: 用户 ID 和角色
        in: body
        name: request
        required: true
        schema:
          properties:
            id:
              type: integer
            role:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg.Response'
      security:
      - BearerAuth: []
      summary: 更新用户角色
      tags:
      - 用户
  /user/upload-avatar:
    post:
      consumes:
      - multipart/form-data
      description: 上传当前用户的头像图片文件
      parameters:
      - description: 头像文件（支持 jpeg、png、gif、webp，最大 5MB）
        in: formData
        name: avatar
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg.Response'
      security:
      - BearerAuth: []
      summary: 上传头像
      tags:
      - 用户
securityDefinitions:
  BearerAuth:
    description: Type "Bearer " followed by your JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"