# app/server/internal/handler/user_handler.go
package handler

import (
	"GOSpeak/internal/pkg"
	"GOSpeak/internal/service"
	"fmt"
	"path/filepath"
	"time"

	"github.com/gin-gonic/gin"
)

type UserHandler struct {
	userService    *service.UserService
	storageService *service.StorageService
}

func NewUserHandler(userService *service.UserService, storageService *service.StorageService) *UserHandler {
	return &UserHandler{userService: userService, storageService: storageService}
}

// GetProfile
// @Summary      获取用户资料
// @Description  获取当前已认证用户的资料信息
// @Tags         用户
// @Produce      json
// @Security     BearerAuth
// @Success      200  {object}  pkg.Response
// @Router       /user/profile [post]
func (h *UserHandler) GetProfile(c *gin.Context) {
	userUUID, ok := c.Get("user_uuid")
	if !ok {
		pkg.Fail(c, pkg.TOKEN_NOT_EXIST)
		return
	}

	uuidStr, ok := userUUID.(string)
	if !ok || uuidStr == "" {
		pkg.Fail(c, pkg.TOKEN_NOT_EXIST)
		return
	}

	user, err := h.userService.GetByUUID(uuidStr)
	if err != nil {
		pkg.HandleError(c, err)
		return
	}

	pkg.Success(c, user)
}

// GetByID
// @Summary      根据 ID 获取用户
// @Description  通过用户 ID 获取用户信息
// @Tags         用户
// @Accept       json
// @Produce      json
// @Security     BearerAuth
// @Param        request  body      object{id=uint}  true  "用户 ID"
// @Success      200  {object}  pkg.Response
// @Router       /user/get [post]
func (h *UserHandler) GetByID(c *gin.Context) {
	var req struct {
		ID uint `json:"id" binding:"required"`
	}
	if err := c.ShouldBindJSON(&req); err != nil {
		pkg.Fail(c, pkg.INVALID_PARAMS, err.Error())
		return
	}

	user, err := h.userService.GetByID(req.ID)
	if err != nil {
		pkg.HandleError(c, err)
		return
	}

	pkg.Success(c, user)
}

// List
// @Summary      获取用户列表
// @Description  分页获取用户列表
// @Tags         用户
// @Accept       json
// @Produce      json
// @Security     BearerAuth
// @Param        request  body      object{page=int,page_size=int}  true  "分页参数"
// @Success      200       {object}  pkg.Response
// @Router       /user/list [post]
func (h *UserHandler) List(c *gin.Context) {
	var req struct {
		Page     int `json:"page"`
		PageSize int `json:"page_size"`
	}
	_ = c.ShouldBindJSON(&req)
	if req.Page <= 0 {
		req.Page = 1
	}
	if req.PageSize <= 0 {
		req.PageSize = 20
	}

	users, total, err := h.userService.List(req.Page, req.PageSize)
	if err != nil {
		pkg.HandleError(c, err)
		return
	}

	pkg.Success(c, gin.H{
		"list":  users,
		"total": total,
		"page":  req.Page,
	})
}

// Delete
// @Summary      删除用户
// @Description  根据用户 ID 删除用户
// @Tags         用户
// @Accept       json
// @Produce      json
// @Security     BearerAuth
// @Param        request  body      object{id=uint}  true  "用户 ID"
// @Success      200  {object}  pkg.Response
// @Router       /user/delete [post]
func (h *UserHandler) Delete(c *gin.Context) {
	var req struct {
		ID uint `json:"id" binding:"required"`
	}
	if err := c.ShouldBindJSON(&req); err != nil {
		pkg.Fail(c, pkg.INVALID_PARAMS, err.Error())
		return
	}

	if err := h.userService.Delete(req.ID); err != nil {
		pkg.HandleError(c, err)
		return
	}

	pkg.Success(c, nil)
}

// GetByName
// @Summary      根据用户名获取用户
// @Description  通过用户名 (identity) 获取用户信息，用于房间成员资料查询
// @Tags         用户
// @Accept       json
// @Produce      json
// @Security     BearerAuth
// @Param        request  body      object{identity=string}  true  "用户名"
// @Success      200  {object}  pkg.Response
// @Router       /user/info [post]
func (h *UserHandler) GetByName(c *gin.Context) {
	var req struct {
		Identity string `json:"identity" binding:"required"`
	}
	if err := c.ShouldBindJSON(&req); err != nil {
		pkg.Fail(c, pkg.INVALID_PARAMS, err.Error())
		return
	}

	user, err := h.userService.GetByName(req.Identity)
	if err != nil {
		pkg.HandleError(c, err)
		return
	}

	pkg.Success(c, user)
}

// UpdateProfile
// @Summary      更新当前用户资料
// @Description  更新当前已认证用户的显示名称和头像
// @Tags         用户
// @Accept       json
// @Produce      json
// @Security     BearerAuth
// @Param        body  body      object  true  "用户资料"
// @Success      200   {object}  pkg.Response
// @Router       /user/update-profile [post]
func (h *UserHandler) UpdateProfile(c *gin.Context) {
	userUUID, ok := c.Get("user_uuid")
	if !ok {
		pkg.Fail(c, pkg.TOKEN_NOT_EXIST)
		return
	}

	uuidStr, ok := userUUID.(string)
	if !ok || uuidStr == "" {
		pkg.Fail(c, pkg.TOKEN_NOT_EXIST)
		return
	}

	var req struct {
		DisplayName string `json:"display_name"`
		Avatar      string `json:"avatar"`
	}
	if err := c.ShouldBindJSON(&req); err != nil {
		pkg.Fail(c, pkg.INVALID_PARAMS, err.Error())
		return
	}

	user, err := h.userService.UpdateProfile(uuidStr, req.DisplayName, req.Avatar)
	if err != nil {
		pkg.HandleError(c, err)
		return
	}

	pkg.Success(c, user)
}

// UploadAvatar
// @Summary      上传头像
// @Description  上传当前用户的头像图片文件
// @Tags         用户
// @Accept       multipart/form-data
// @Produce      json
// @Security     BearerAuth
// @Param        avatar  formData  file  true  "头像文件（支持 jpeg、png、gif、webp，最大 5MB）"
// @Success      200     {object}  pkg.Response
// @Router       /user/upload-avatar [post]
func (h *UserHandler) UploadAvatar(c *gin.Context) {
	userUUID, ok := c.Get("user_uuid")
	if !ok {
		pkg.Fail(c, pkg.TOKEN_NOT_EXIST)
		return
	}

	uuidStr, ok := userUUID.(string)
	if !ok || uuidStr == "" {
		pkg.Fail(c, pkg.TOKEN_NOT_EXIST)
		return
	}

	file, err := c.FormFile("avatar")
	if err != nil {
		pkg.Fail(c, pkg.INVALID_PARAMS, "avatar file is required")
		return
	}

	// 从存储配置获取大小限制
	cfg, err := h.storageService.GetConfig()
	if err != nil {
		pkg.HandleError(c, err)
		return
	}

	// 文件大小校验
	maxBytes := int64(cfg.MaxFileSize) * 1024 * 1024
	if file.Size > maxBytes {
		pkg.Fail(c, pkg.STORAGE_FILE_TOO_LARGE)
		return
	}

	// 文件类型校验
	contentType := file.Header.Get("Content-Type")
	allowedTypes := map[string]bool{
		"image/jpeg": true,
		"image/png":  true,
		"image/gif":  true,
		"image/webp": true,
	}
	if !allowedTypes[contentType] {
		pkg.Fail(c, pkg.INVALID_PARAMS, "invalid image type, allowed: jpeg, png, gif, webp")
		return
	}

	// 生成对象键: avatars/{uuid}_{timestamp}.{ext}
	ext := filepath.Ext(file.Filename)
	if ext == "" {
		switch contentType {
		case "image/jpeg":
			ext = ".jpg"
		case "image/png":
			ext = ".png"
		case "image/gif":
			ext = ".gif"
		case "image/webp":
			ext = ".webp"
		default:
			ext = ".bin"
		}
	}
	objectKey := fmt.Sprintf("avatars/%s_%d%s", uuidStr, time.Now().UnixMilli(), ext)

	// 通过存储抽象层上传（S3 或本地）
	src, err := file.Open()
	if err != nil {
		pkg.Fail(c, pkg.INTERNAL_ERROR, "failed to open uploaded file")
		return
	}
	defer src.Close()

	avatarURL, err := h.storageService.UploadFromReader(objectKey, src, file.Size, contentType)
	if err != nil {
		pkg.HandleError(c, err)
		return
	}

	pkg.Success(c, gin.H{"avatar": avatarURL})
}

// UpdateRole
// @Summary      更新用户角色
// @Description  管理员更新用户角色（admin/user）
// @Tags         用户
// @Accept       json
// @Produce      json
// @Security     BearerAuth
// @Param        request  body      object{id=uint,role=string}  true  "用户 ID 和角色"
// @Success      200   {object}  pkg.Response
// @Router       /user/update-role [post]
func (h *UserHandler) UpdateRole(c *gin.Context) {
	var req struct {
		ID   uint   `json:"id" binding:"required"`
		Role string `json:"role" binding:"required"`
	}
	if err := c.ShouldBindJSON(&req); err != nil {
		pkg.Fail(c, pkg.INVALID_PARAMS, err.Error())
		return
	}

	if err := h.userService.UpdateRole(req.ID, req.Role); err != nil {
		pkg.HandleError(c, err)
		return
	}

	pkg.Success(c, nil)
}